+++ tests/Cashout/InitializerTest.php
@@ -6,6 +6,7 @@ use DateTime;
 use HiPay\Wallet\Mirakl\Cashout\Initializer;
 use HiPay\Wallet\Mirakl\Cashout\Model\Operation\OperationInterface;
 use HiPay\Wallet\Mirakl\Test\Common\AbstractProcessorTest;
+use HiPay\Wallet\Mirakl\Test\Stub\Entity\LogOperations;
 use HiPay\Wallet\Mirakl\Test\Stub\Api\Mirakl;
 use HiPay\Wallet\Mirakl\Test\Stub\Entity\Operation;
 use HiPay\Wallet\Mirakl\Test\Stub\Entity\Vendor;
@@ -13,6 +14,7 @@ use HiPay\Wallet\Mirakl\Vendor\Model\VendorInterface;
 use HiPay\Wallet\Mirakl\Vendor\Model\VendorManagerInterface;
 use Prophecy\Argument;
 use Prophecy\Argument\Token\TypeToken;
+use HiPay\Wallet\Mirakl\Cashout\Model\Operation\Status;
 
 /**
  *
@@ -25,6 +27,9 @@ class InitializerTest extends AbstractProcessorTest
 {
     /** @var  VendorInterface */
     protected $technicalAccountArgument;
+
+    /** @var Transfer  */
+    protected $transferArgument;
 
     /** @var  OperationInterface|TypeToken */
     private $operationArgument;
@@ -41,6 +46,8 @@ class InitializerTest extends AbstractProcessorTest
             Argument::type("\\HiPay\\Wallet\\Mirakl\\Cashout\\Model\\Operation\\OperationInterface");
 
         $this->vendorArgument = Argument::type("\\HiPay\\Wallet\\Mirakl\\Vendor\\Model\\VendorInterface");
+
+        $this->transferArgument = Argument::type("\\HiPay\\Wallet\\Mirakl\\Api\\HiPay\\Model\\Soap\\Transfer");
 
         /** @var VendorInterface vendorArgument */
         $this->technicalAccountArgument = Argument::is($this->technical);
@@ -208,6 +215,17 @@ class InitializerTest extends AbstractProcessorTest
         $this->operationManager->findByMiraklIdAndPaymentVoucherNumber(Argument::is(false), Argument::type("string"))
             ->willReturn(null)
             ->shouldBeCalled();
+
+        $this->operationManager->findByStatus(Argument::type("HiPay\Wallet\Mirakl\Cashout\Model\Operation\Status"))
+            ->willReturn(array())
+            ->shouldBeCalled();
+
+        $this->operationManager->findByStatusAndBeforeUpdatedAt(Argument::type("HiPay\Wallet\Mirakl\Cashout\Model\Operation\Status"), Argument::type("Datetime"))
+            ->willReturn(array())
+            ->shouldBeCalled();
+
+//        $this->cashoutInitializer->transferOperations()
+//            ->willReturn()->shouldBeCalled();
 
         $this->operationManager->isValid($this->operationArgument)->willReturn(true)->shouldBeCalled();
 
@@ -552,5 +570,91 @@ class InitializerTest extends AbstractProcessorTest
                 return new Operation($amount, $cycleDate, $paymentVoucher, $miraklId);
             })->shouldBeCalled();
         }
+    }
+
+    /**
+     * @cover ::transfer
+     * @group transfer
+     */
+    public function testVendorTransferSuccessful()
+    {
+        $transferId = rand();
+        $this->vendorManager->findByMiraklId(Argument::type("integer"))
+                            ->willReturn(new Vendor("test@test.com", rand(), rand()))
+                            ->shouldBeCalled();
+
+        $this->hipay->isAvailable(Argument::containingString("@"), Argument::any())
+                    ->willReturn(false)
+                    ->shouldBeCalled();
+
+        $this->hipay->transfer($this->transferArgument, Argument::cetera())
+                    ->willReturn($transferId)
+                    ->shouldBeCalled();
+
+        $operation = new Operation(2000, new DateTime(), "000001", rand());
+
+        $this->logOperationsManager->save(Argument::any())->willReturn()->shouldBeCalled();
+
+        $this->logOperationsManager->findByMiraklIdAndPaymentVoucherNumber(Argument::any(), Argument::any())->willReturn(new LogOperations(200, 2001))->shouldBeCalled();
+
+        $result = $this->cashoutInitializer->transfer($operation);
+
+        $this->assertInternalType("integer", $result);
+
+        $this->assertEquals($transferId, $result);
+
+        $this->assertEquals(Status::TRANSFER_SUCCESS, $operation->getStatus());
+    }
+
+    /**
+     * @cover ::transfer
+     * @group transfer
+     */
+    public function testOperatorTransferSuccessful()
+    {
+        $transferId = rand();
+        $operation = new Operation(2000, new DateTime(), "000001", false);
+
+        $this->hipay->isAvailable(Argument::containingString("@"), Argument::any())->willReturn(false)->shouldBeCalled();
+        $this->hipay->transfer($this->transferArgument, Argument::cetera())->willReturn($transferId)->shouldBeCalled();
+
+        $this->logOperationsManager->save(Argument::any())->willReturn()->shouldBeCalled();
+
+        $this->logOperationsManager->findByMiraklIdAndPaymentVoucherNumber(Argument::any(), Argument::any())->willReturn(new LogOperations(200, 2001))->shouldBeCalled();
+
+        $result = $this->cashoutInitializer->transfer($operation);
+
+        $this->assertInternalType("integer", $result);
+
+        $this->assertEquals($transferId, $result);
+
+        $this->assertEquals(Status::TRANSFER_SUCCESS, $operation->getStatus());
+    }
+
+    /**
+     * @cover ::transfer
+     * @group transfer
+     */
+    public function testTransferWalletNotFound()
+    {
+        $operation = new Operation(2000, new DateTime(), "000001", rand());
+
+        $this->hipay->isAvailable(Argument::containingString("@"), Argument::any())->willReturn(true)->shouldBeCalled();
+
+        $this->hipay->transfer(Argument::any())->shouldNotBeCalled();
+
+        $this->vendorManager->findByMiraklId(Argument::type("integer"))
+                            ->willReturn(new Vendor("test@test.com", rand(), rand()))
+                            ->shouldBeCalled();
+
+        $this->logOperationsManager->save(Argument::any())->willReturn()->shouldBeCalled();
+
+        $this->logOperationsManager->findByMiraklIdAndPaymentVoucherNumber(Argument::any(), Argument::any())->willReturn(new LogOperations(200, 2001))->shouldBeCalled();
+
+        $this->setExpectedException("\\HiPay\\Wallet\\Mirakl\\Exception\\WalletNotFoundException");
+
+        $this->cashoutInitializer->transfer($operation);
+
+        $this->assertEquals(Status::TRANSFER_FAILED, $operation->getStatus());
     }
 }
